[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 78.2 MB -BaekjoonHub

+++ 프로그래머스/lv0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -3,7 +3,7 @@ class Solution {
         int answer = 0;
         
         if (0 <= num1 && num1 <= 100 && 0 <= num2 && num2 <= 100) {
-            float a = (float) num1 / num2 * 1000;
+            double a = (double) num1 / num2 * 1000;
             answer = (int) a;
         }
         return answer;
